add JPetOptionsGenerator to create options from varioable_maps insted of JPetCmdParser
