+++ JPetCmdParser/JPetCmdParser.h
@@ -35,71 +35,6 @@ public:
   inline const po::options_description getOptionsDescription() const {
     return fOptionsDescriptions;
   }
-  std::vector<JPetOptions> generateOptions(const po::variables_map& optsMap, const std::map<std::string, std::string>& additionalOptions = {}) const;
-
-  bool areCorrectOptions(const po::variables_map& options) const;
-  inline const std::vector<std::string>& getFileNames(const po::variables_map& variablesMap) const {
-    return variablesMap["file"].as< std::vector<std::string> >();
-  }
-  inline bool isCorrectFileType(const std::string& type) const {
-    if (type == "hld" || type == "root" || type == "scope" || type == "zip") {
-      return true;
-    }
-    return false;
-  }
-
-  inline const std::string& getOutputPath(const po::variables_map& variablesMap) const {
-    return variablesMap["outputPath"].as<std::string>();
-  }
-
-  inline bool isOutputPath(const po::variables_map& variablesMap) const {
-    return (bool)variablesMap.count("outputPath");
-  }
-
-  inline const std::string& getFileType(const po::variables_map& variablesMap) const {
-    return variablesMap["type"].as<std::string>();
-  }
-
-  inline bool IsFileTypeSet(const po::variables_map& variablesMap) const {
-    return (bool)variablesMap.count("type");
-  }
-  inline int getLowerEventBound(const po::variables_map& variablesMap) const {
-    return variablesMap["range"].as< std::vector<int> >()[0];
-  }
-  inline int getHigherEventBound(const po::variables_map& variablesMap) const {
-    return variablesMap["range"].as< std::vector<int> >()[1];
-  }
-  inline bool isParamSet(const po::variables_map& variablesMap) const {
-    return (bool)variablesMap.count("param");
-  }
-  inline const std::string& getParam(const po::variables_map& variablesMap) const {
-    return variablesMap["param"].as< std::string >();
-  }
-
-  inline bool isRunNumberSet(const po::variables_map& variablesMap) const {
-    return (bool)variablesMap.count("runId");
-  }
-  inline const int getRunNumber(const po::variables_map& variablesMap) const {
-    return variablesMap["runId"].as<int>();
-  }
-
-  inline bool isProgressBarSet(const po::variables_map& variablesMap) const {
-    return variablesMap["progressBar"].as<bool>();
-  }
-
-  static inline bool isLocalDBSet(const po::variables_map& variablesMap) {
-    return variablesMap.count("localDB") > 0;
-  }
-  static inline std::string getLocalDBName(const po::variables_map& variablesMap) {
-    return variablesMap["localDB"].as<std::string>();
-  }
-
-  static inline bool isLocalDBCreateSet(const po::variables_map& variablesMap) {
-    return variablesMap.count("localDBCreate") > 0;
-  }
-  static inline std::string getLocalDBCreateName(const po::variables_map& variablesMap) {
-    return variablesMap["localDBCreate"].as<std::string>();
-  }
 
 protected:
   po::options_description fOptionsDescriptions;
