+++ JPetOptionsGenerator/JPetOptionsGeneratorTest.cpp
@@ -1,0 +1,23 @@
+/**
+ *  @copyright Copyright 2016 The J-PET Framework Authors. All rights reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may find a copy of the License in the LICENCE file.
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ *  @file JPetOptionsGeneratorTest.cpp
+ */
+
+#define BOOST_TEST_DYN_LINK
+#define BOOST_TEST_MODULE JPetOptionsGeneratorTest
+#include <boost/test/unit_test.hpp>
+#include <cstdlib>
+#include "../JPetCmdParser/JPetCmdParser.h"
+#include "../JPetOptionsGenerator/JPetOptionsGenerator.h"
+
+using namespace std;
